+++ 11/02/src/test/java/tests/Config.java
@@ -1,5 +1,5 @@
 package tests;
 
-public interface Config {
-    final String baseUrl = System.getProperty("baseUrl", "http://the-internet.herokuapp.com");
+public class Config {
+    public static final String baseUrl = System.getProperty("baseUrl", "http://the-internet.herokuapp.com");
 }