+++ 12/src/test/java/tests/Base.java
@@ -5,8 +5,9 @@ import org.junit.rules.ExternalResource;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
+import static tests.Config.*;
 
-public class Base implements Config {
+public class Base {
 
     protected WebDriver driver;