+++ 12/src/test/java/tests/Config.java
@@ -1,6 +1,6 @@
 package tests;
 
-public interface Config {
-    final String baseUrl = System.getProperty("baseUrl", "http://the-internet.herokuapp.com");
-    final String browser = System.getProperty("browser", "firefox");
+public class Config {
+    public static final String baseUrl = System.getProperty("baseUrl", "http://the-internet.herokuapp.com");
+    public static final String browser = System.getProperty("browser", "firefox");
 }