+++ 16/03/src/test/java/tests/TestLogin.java
@@ -3,8 +3,8 @@ package tests;
 import org.junit.Test;
 import org.junit.Before;
 import static org.junit.Assert.*;
-import pageobjects.Login;
 import org.junit.experimental.categories.Category;
+import pageobjects.Login;
 import tests.groups.Shallow;
 
 public class TestLogin extends Base {