+++ changelog.md
@@ -1,8 +1,9 @@
 # 1.0.1
 ## November 2015
-- Chapter 13, Part 2. Added missing @Override to the starting method in tests/Base.java (Mark Winspear)
+- 13/02 thru 16/03. Added missing @Override to the starting method in tests/Base.java (Mark Winspear)
 - Updated Selenium from 2.45.0 to 2.48.2
 - Updated JUnit from 4.11 to 4.12
+- 11/02 thru 16/03. Ported Config.java from an interface to a static class (Benjamin Limb)
 
 # 1.0.0
 ## March 2015